+++ game.c
@@ -4,9 +4,10 @@
 #include <stdlib.h>
 #include <math.h>
 
+#include <gtk/gtk.h>//for debug
 #include "game.h"
 
-void tensor(const double a[MATRIX_DIM][MATRIX_DIM], const double b[MATRIX_DIM][MATRIX_DIM], double c[VECTOR_DIM][VECTOR_DIM])
+void tensor(const long double a[MATRIX_DIM][MATRIX_DIM], const long double b[MATRIX_DIM][MATRIX_DIM], long double c[VECTOR_DIM][VECTOR_DIM])
 //Computes the tensor product of two matrices of dimension MATRIX_DIM.
 {
 	for(int i = 0 ; i < VECTOR_DIM ; i++)
@@ -16,10 +17,10 @@ void tensor(const double a[MATRIX_DIM][MATRIX_DIM], const double b[MATRIX_DIM][M
 	}
 }
 
-Point matrixVectorProduct(const double m[VECTOR_DIM][VECTOR_DIM], const Point *vector)
+Point matrixVectorProduct(const long double m[VECTOR_DIM][VECTOR_DIM], const Point *vector)
 {
 	unsigned int i = 0, j = 0;
-	double lines[VECTOR_DIM] = {0}, vect2List[VECTOR_DIM] = {0};
+	long double lines[VECTOR_DIM] = {0}, vect2List[VECTOR_DIM] = {0};
 	Point result;
 
 	vect2List[0] = vector->x;
@@ -41,12 +42,12 @@ Point matrixVectorProduct(const double m[VECTOR_DIM][VECTOR_DIM], const Point *v
 	return result;
 }
 
-double norm2(const Point *point)
+long double norm2(const Point *point)
 {
 	return sqrt(point->x * point->x + point->y * point->y + point->z * point->z + point->h * point->h);
 }
 
-double norm2_3D(const Point *point)
+long double norm2_3D(const Point *point)
 {
 	return sqrt(point->x * point->x + point->y * point->y + point->z * point->z);
 }
@@ -54,7 +55,7 @@ double norm2_3D(const Point *point)
 void normalize(Point *point)
 {
 	//We call "alpha" the normalization coef.
-	double alpha = norm2(point);
+	long double alpha = norm2(point);
 	point->x /= alpha;
 	point->y /= alpha;
 	point->z /= alpha;
@@ -64,17 +65,17 @@ void normalize(Point *point)
 void normalize_3D(Point *point)
 {//We only normalize the 3D part of a 4D vector.
 	//We call "alpha" the normalization coef.
-	double alpha = norm2_3D(point);
+	long double alpha = norm2_3D(point);
 	point->x /= alpha;
 	point->y /= alpha;
 	point->z /= alpha;
 }
 
-double proba(const Point *psi, const double m[MATRIX_DIM][MATRIX_DIM], const double n[MATRIX_DIM][MATRIX_DIM])
+long double proba(const Point *psi, const long double m[MATRIX_DIM][MATRIX_DIM], const long double n[MATRIX_DIM][MATRIX_DIM])
 //Returns the probability of Alice answering "a" and Bob answering "b" given the setting (psi,m,n) (where m -> a, n -> b)
 {
 	//a0 is the first column of m, a1 the second.
-	double a0[2], a1[2], b0[2], b1[2];
+	long double a0[2], a1[2], b0[2], b1[2];
 	a0[0] = m[0][0];
 	a0[1] = m[1][0];
 	a1[0] = m[0][1];
@@ -88,17 +89,17 @@ double proba(const Point *psi, const double m[MATRIX_DIM][MATRIX_DIM], const dou
 	return psi-> x * psi->x * NORM2D2(a0) * NORM2D2(b0) + psi->y * psi->y * NORM2D2(a0) * NORM2D2(b1) + psi->z * psi->z * NORM2D2(a1) * NORM2D2(b0) + psi->h * psi->h * NORM2D2(a1) * NORM2D2(b1) + 2 * psi->x * psi->y * NORM2D2(a0) * SCALAR(b0,b1) + 2 * psi->x * psi->z * NORM2D2(b0) * SCALAR(a0,a1) + 2 * psi->x * psi->h * SCALAR(a0,a1) * SCALAR(b0,b1) + 2 * psi->y * psi->z * SCALAR(a0,a1) * SCALAR(b0,b1) + 2 * psi->y * psi->h * SCALAR(a0,a1) * NORM2D2(b1) + 2 * psi->z * psi->h * NORM2D2(a1) * SCALAR(b0,b1);
 }
 
-Point measure(const Point *psi, const double m[MATRIX_DIM][MATRIX_DIM], const double n[MATRIX_DIM][MATRIX_DIM])
+Point measure(const Point *psi, const long double m[MATRIX_DIM][MATRIX_DIM], const long double n[MATRIX_DIM][MATRIX_DIM])
 {//This function returns the post measurement state after measuring with m and n.
 	Point postState;
-	double *c = NULL;
+	long double *c = NULL;
 
-	c = (double*)malloc(VECTOR_DIM * VECTOR_DIM * sizeof(double));
+	c = (long double*)malloc(VECTOR_DIM * VECTOR_DIM * sizeof(long double));
 	if(c == NULL)
 		exit(EXIT_FAILURE);
 
-	tensor(m,n,(double (*)[VECTOR_DIM])c);
-	postState = matrixVectorProduct((double (*)[VECTOR_DIM])c, psi);
+	tensor(m,n,(long double (*)[VECTOR_DIM])c);
+	postState = matrixVectorProduct((long double (*)[VECTOR_DIM])c, psi);
 
 	free(c);
 	normalize(&postState);
