+++ game.h
@@ -6,13 +6,13 @@
 
 #include "Structs.h"
 
-void tensor(const double a[MATRIX_DIM][MATRIX_DIM], const double b[MATRIX_DIM][MATRIX_DIM], double c[VECTOR_DIM][VECTOR_DIM]);
-double norm2(const Point *point);
-double norm2_3D(const Point *point);
+void tensor(const long double a[MATRIX_DIM][MATRIX_DIM], const long double b[MATRIX_DIM][MATRIX_DIM], long double c[VECTOR_DIM][VECTOR_DIM]);
+long double norm2(const Point *point);
+long double norm2_3D(const Point *point);
 void normalize(Point *point);
 void normalize_3D(Point *point);
-double proba(const Point *psi, const double m[MATRIX_DIM][MATRIX_DIM], const double n[MATRIX_DIM][MATRIX_DIM]);
-Point measure(const Point *psi, const double m[MATRIX_DIM][MATRIX_DIM], const double n[MATRIX_DIM][MATRIX_DIM]);
-Point matrixVectorProduct(const double m[VECTOR_DIM][VECTOR_DIM], const Point *vector);
+long double proba(const Point *psi, const long double m[MATRIX_DIM][MATRIX_DIM], const long double n[MATRIX_DIM][MATRIX_DIM]);
+Point measure(const Point *psi, const long double m[MATRIX_DIM][MATRIX_DIM], const long double n[MATRIX_DIM][MATRIX_DIM]);
+Point matrixVectorProduct(const long double m[VECTOR_DIM][VECTOR_DIM], const Point *vector);
 
 #endif
